FixBug : fixed where search for remittance by remittance number was doing the search in the current page list , when data for remittance got over a billion overflow exceptions came through so i changed the integers to long

+++ FreightAccounting.Core/Entities/Debtor.cs
@@ -23,7 +23,7 @@ public class Debtor : BaseEntity
     /// <summary>
     /// مقدار بدهی
     /// </summary>
-    public int DebtAmount { get; set; }
+    public long DebtAmount { get; set; }
 
     /// <summary>
     /// تاریخ پرداخت میتونه نال باشه
